requirements.txt
Flask==3.0.2
Pillow==10.2.0
pyTelegramBotAPI==4.15.4
qrcode==7.4.2
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.31.0
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pypng==0.20220715.0
pymongo[srv]==4.6.1
